# Remove LatestMap, to reduce memory allocation