--- a/report/latest_map_generated.go
+++ b/report/latest_map_generated.go
@@ -1,37 +1,36 @@
 // Generated file, do not edit.
-// To regenerate, run ./tools/generate_latest_map ./report/latest_map_generated.go string NodeControlData
+// To regenerate, run ../extras/generate_latest_map ./latest_map_generated.go string NodeControlData
 
 package report
 
 import (
+	"fmt"
 	"time"
 
 	"github.com/ugorji/go/codec"
+	"github.com/weaveworks/ps"
 )
 
-type wireStringLatestEntry struct {
+type stringLatestEntry struct {
 	Timestamp time.Time `json:"timestamp"`
 	Value     string    `json:"value"`
 }
 
-type stringLatestEntryDecoder struct{}
-
-func (d *stringLatestEntryDecoder) Decode(decoder *codec.Decoder, entry *LatestEntry) {
-	wire := wireStringLatestEntry{}
-	decoder.Decode(&wire)
-	entry.Timestamp = wire.Timestamp
-	entry.Value = wire.Value
+// String returns the StringLatestEntry's string representation.
+func (e *stringLatestEntry) String() string {
+	return fmt.Sprintf("%v (%s)", e.Value, e.Timestamp.String())
 }
 
-// StringLatestEntryDecoder is an implementation of LatestEntryDecoder
-// that decodes the LatestEntry instances having a string value.
-var StringLatestEntryDecoder LatestEntryDecoder = &stringLatestEntryDecoder{}
+// Equal returns true if the supplied StringLatestEntry is equal to this one.
+func (e *stringLatestEntry) Equal(e2 *stringLatestEntry) bool {
+	return e.Timestamp.Equal(e2.Timestamp) && e.Value == e2.Value
+}
 
 // StringLatestMap holds latest string instances.
-type StringLatestMap LatestMap
+type StringLatestMap struct{ ps.Map }
 
 // EmptyStringLatestMap is an empty StringLatestMap. Start with this.
-var EmptyStringLatestMap = (StringLatestMap)(MakeLatestMapWithDecoder(StringLatestEntryDecoder))
+var EmptyStringLatestMap = StringLatestMap{ps.NewMap()}
 
 // MakeStringLatestMap makes an empty StringLatestMap.
 func MakeStringLatestMap() StringLatestMap {
@@ -40,77 +39,117 @@ func MakeStringLatestMap() StringLatestMap {
 
 // Copy is a noop, as StringLatestMaps are immutable.
 func (m StringLatestMap) Copy() StringLatestMap {
-	return (StringLatestMap)((LatestMap)(m).Copy())
+	return m
 }
 
 // Size returns the number of elements.
 func (m StringLatestMap) Size() int {
-	return (LatestMap)(m).Size()
+	if m.Map == nil {
+		return 0
+	}
+	return m.Map.Size()
 }
 
 // Merge produces a fresh StringLatestMap containing the keys from both inputs.
 // When both inputs contain the same key, the newer value is used.
 func (m StringLatestMap) Merge(other StringLatestMap) StringLatestMap {
-	return (StringLatestMap)((LatestMap)(m).Merge((LatestMap)(other)))
+	output := mergeMaps(m.Map, other.Map, func(a, b interface{}) bool {
+		return a.(*stringLatestEntry).Timestamp.Before(b.(*stringLatestEntry).Timestamp)
+	})
+	return StringLatestMap{output}
 }
 
 // Lookup the value for the given key.
 func (m StringLatestMap) Lookup(key string) (string, bool) {
-	v, ok := (LatestMap)(m).Lookup(key)
+	v, _, ok := m.LookupEntry(key)
 	if !ok {
 		var zero string
 		return zero, false
 	}
-	return v.(string), true
+	return v, true
 }
 
 // LookupEntry returns the raw entry for the given key.
 func (m StringLatestMap) LookupEntry(key string) (string, time.Time, bool) {
-	v, timestamp, ok := (LatestMap)(m).LookupEntry(key)
+	if m.Map == nil {
+		var zero string
+		return zero, time.Time{}, false
+	}
+	value, ok := m.Map.Lookup(key)
 	if !ok {
 		var zero string
-		return zero, timestamp, false
+		return zero, time.Time{}, false
 	}
-	return v.(string), timestamp, true
+	e := value.(*stringLatestEntry)
+	return e.Value, e.Timestamp, true
 }
 
 // Set the value for the given key.
 func (m StringLatestMap) Set(key string, timestamp time.Time, value string) StringLatestMap {
-	return (StringLatestMap)((LatestMap)(m).Set(key, timestamp, value))
+	if m.Map == nil {
+		m.Map = ps.NewMap()
+	}
+	return StringLatestMap{m.Map.Set(key, &stringLatestEntry{timestamp, value})}
 }
 
 // Delete the value for the given key.
 func (m StringLatestMap) Delete(key string) StringLatestMap {
-	return (StringLatestMap)((LatestMap)(m).Delete(key))
+	if m.Map == nil {
+		return m
+	}
+	return StringLatestMap{m.Map.Delete(key)}
 }
 
 // ForEach executes fn on each key value pair in the map.
 func (m StringLatestMap) ForEach(fn func(k string, timestamp time.Time, v string)) {
-	(LatestMap)(m).ForEach(func(key string, ts time.Time, value interface{}) {
-		fn(key, ts, value.(string))
-	})
+	if m.Map != nil {
+		m.Map.ForEach(func(key string, value interface{}) {
+			fn(key, value.(*stringLatestEntry).Timestamp, value.(*stringLatestEntry).Value)
+		})
+	}
 }
 
 // String returns the StringLatestMap's string representation.
 func (m StringLatestMap) String() string {
-	return (LatestMap)(m).String()
+	return mapToString(m.Map)
 }
 
 // DeepEqual tests equality with other StringLatestMap.
 func (m StringLatestMap) DeepEqual(n StringLatestMap) bool {
-	return (LatestMap)(m).DeepEqual((LatestMap)(n))
+	return mapEqual(m.Map, n.Map, func(val, otherValue interface{}) bool {
+		return val.(*stringLatestEntry).Equal(otherValue.(*stringLatestEntry))
+	})
+}
+
+func (m StringLatestMap) toIntermediate() map[string]stringLatestEntry {
+	intermediate := make(map[string]stringLatestEntry, m.Size())
+	if m.Map != nil {
+		m.Map.ForEach(func(key string, val interface{}) {
+			intermediate[key] = *val.(*stringLatestEntry)
+		})
+	}
+	return intermediate
 }
 
 // CodecEncodeSelf implements codec.Selfer.
 func (m *StringLatestMap) CodecEncodeSelf(encoder *codec.Encoder) {
-	(*LatestMap)(m).CodecEncodeSelf(encoder)
+	if m.Map != nil {
+		encoder.Encode(m.toIntermediate())
+	} else {
+		encoder.Encode(nil)
+	}
 }
 
 // CodecDecodeSelf implements codec.Selfer.
 func (m *StringLatestMap) CodecDecodeSelf(decoder *codec.Decoder) {
-	bm := (*LatestMap)(m)
-	bm.decoder = StringLatestEntryDecoder
-	bm.CodecDecodeSelf(decoder)
+	out := mapRead(decoder, func(isNil bool) interface{} {
+		value := &stringLatestEntry{}
+		if !isNil {
+			decoder.Decode(value)
+		}
+		return value
+	})
+	*m = StringLatestMap{out}
 }
 
 // MarshalJSON shouldn't be used, use CodecEncodeSelf instead.
@@ -123,29 +162,26 @@ func (*StringLatestMap) UnmarshalJSON(b []byte) error {
 	panic("UnmarshalJSON shouldn't be used, use CodecDecodeSelf instead")
 }
 
-type wireNodeControlDataLatestEntry struct {
+type nodeControlDataLatestEntry struct {
 	Timestamp time.Time       `json:"timestamp"`
 	Value     NodeControlData `json:"value"`
 }
 
-type nodeControlDataLatestEntryDecoder struct{}
-
-func (d *nodeControlDataLatestEntryDecoder) Decode(decoder *codec.Decoder, entry *LatestEntry) {
-	wire := wireNodeControlDataLatestEntry{}
-	decoder.Decode(&wire)
-	entry.Timestamp = wire.Timestamp
-	entry.Value = wire.Value
+// String returns the StringLatestEntry's string representation.
+func (e *nodeControlDataLatestEntry) String() string {
+	return fmt.Sprintf("%v (%s)", e.Value, e.Timestamp.String())
 }
 
-// NodeControlDataLatestEntryDecoder is an implementation of LatestEntryDecoder
-// that decodes the LatestEntry instances having a NodeControlData value.
-var NodeControlDataLatestEntryDecoder LatestEntryDecoder = &nodeControlDataLatestEntryDecoder{}
+// Equal returns true if the supplied StringLatestEntry is equal to this one.
+func (e *nodeControlDataLatestEntry) Equal(e2 *nodeControlDataLatestEntry) bool {
+	return e.Timestamp.Equal(e2.Timestamp) && e.Value == e2.Value
+}
 
 // NodeControlDataLatestMap holds latest NodeControlData instances.
-type NodeControlDataLatestMap LatestMap
+type NodeControlDataLatestMap struct{ ps.Map }
 
 // EmptyNodeControlDataLatestMap is an empty NodeControlDataLatestMap. Start with this.
-var EmptyNodeControlDataLatestMap = (NodeControlDataLatestMap)(MakeLatestMapWithDecoder(NodeControlDataLatestEntryDecoder))
+var EmptyNodeControlDataLatestMap = NodeControlDataLatestMap{ps.NewMap()}
 
 // MakeNodeControlDataLatestMap makes an empty NodeControlDataLatestMap.
 func MakeNodeControlDataLatestMap() NodeControlDataLatestMap {
@@ -154,77 +190,117 @@ func MakeNodeControlDataLatestMap() NodeControlDataLatestMap {
 
 // Copy is a noop, as NodeControlDataLatestMaps are immutable.
 func (m NodeControlDataLatestMap) Copy() NodeControlDataLatestMap {
-	return (NodeControlDataLatestMap)((LatestMap)(m).Copy())
+	return m
 }
 
 // Size returns the number of elements.
 func (m NodeControlDataLatestMap) Size() int {
-	return (LatestMap)(m).Size()
+	if m.Map == nil {
+		return 0
+	}
+	return m.Map.Size()
 }
 
 // Merge produces a fresh NodeControlDataLatestMap containing the keys from both inputs.
 // When both inputs contain the same key, the newer value is used.
 func (m NodeControlDataLatestMap) Merge(other NodeControlDataLatestMap) NodeControlDataLatestMap {
-	return (NodeControlDataLatestMap)((LatestMap)(m).Merge((LatestMap)(other)))
+	output := mergeMaps(m.Map, other.Map, func(a, b interface{}) bool {
+		return a.(*nodeControlDataLatestEntry).Timestamp.Before(b.(*nodeControlDataLatestEntry).Timestamp)
+	})
+	return NodeControlDataLatestMap{output}
 }
 
 // Lookup the value for the given key.
 func (m NodeControlDataLatestMap) Lookup(key string) (NodeControlData, bool) {
-	v, ok := (LatestMap)(m).Lookup(key)
+	v, _, ok := m.LookupEntry(key)
 	if !ok {
 		var zero NodeControlData
 		return zero, false
 	}
-	return v.(NodeControlData), true
+	return v, true
 }
 
 // LookupEntry returns the raw entry for the given key.
 func (m NodeControlDataLatestMap) LookupEntry(key string) (NodeControlData, time.Time, bool) {
-	v, timestamp, ok := (LatestMap)(m).LookupEntry(key)
+	if m.Map == nil {
+		var zero NodeControlData
+		return zero, time.Time{}, false
+	}
+	value, ok := m.Map.Lookup(key)
 	if !ok {
 		var zero NodeControlData
-		return zero, timestamp, false
+		return zero, time.Time{}, false
 	}
-	return v.(NodeControlData), timestamp, true
+	e := value.(*nodeControlDataLatestEntry)
+	return e.Value, e.Timestamp, true
 }
 
 // Set the value for the given key.
 func (m NodeControlDataLatestMap) Set(key string, timestamp time.Time, value NodeControlData) NodeControlDataLatestMap {
-	return (NodeControlDataLatestMap)((LatestMap)(m).Set(key, timestamp, value))
+	if m.Map == nil {
+		m.Map = ps.NewMap()
+	}
+	return NodeControlDataLatestMap{m.Map.Set(key, &nodeControlDataLatestEntry{timestamp, value})}
 }
 
 // Delete the value for the given key.
 func (m NodeControlDataLatestMap) Delete(key string) NodeControlDataLatestMap {
-	return (NodeControlDataLatestMap)((LatestMap)(m).Delete(key))
+	if m.Map == nil {
+		return m
+	}
+	return NodeControlDataLatestMap{m.Map.Delete(key)}
 }
 
 // ForEach executes fn on each key value pair in the map.
 func (m NodeControlDataLatestMap) ForEach(fn func(k string, timestamp time.Time, v NodeControlData)) {
-	(LatestMap)(m).ForEach(func(key string, ts time.Time, value interface{}) {
-		fn(key, ts, value.(NodeControlData))
-	})
+	if m.Map != nil {
+		m.Map.ForEach(func(key string, value interface{}) {
+			fn(key, value.(*nodeControlDataLatestEntry).Timestamp, value.(*nodeControlDataLatestEntry).Value)
+		})
+	}
 }
 
 // String returns the NodeControlDataLatestMap's string representation.
 func (m NodeControlDataLatestMap) String() string {
-	return (LatestMap)(m).String()
+	return mapToString(m.Map)
 }
 
 // DeepEqual tests equality with other NodeControlDataLatestMap.
 func (m NodeControlDataLatestMap) DeepEqual(n NodeControlDataLatestMap) bool {
-	return (LatestMap)(m).DeepEqual((LatestMap)(n))
+	return mapEqual(m.Map, n.Map, func(val, otherValue interface{}) bool {
+		return val.(*nodeControlDataLatestEntry).Equal(otherValue.(*nodeControlDataLatestEntry))
+	})
+}
+
+func (m NodeControlDataLatestMap) toIntermediate() map[string]nodeControlDataLatestEntry {
+	intermediate := make(map[string]nodeControlDataLatestEntry, m.Size())
+	if m.Map != nil {
+		m.Map.ForEach(func(key string, val interface{}) {
+			intermediate[key] = *val.(*nodeControlDataLatestEntry)
+		})
+	}
+	return intermediate
 }
 
 // CodecEncodeSelf implements codec.Selfer.
 func (m *NodeControlDataLatestMap) CodecEncodeSelf(encoder *codec.Encoder) {
-	(*LatestMap)(m).CodecEncodeSelf(encoder)
+	if m.Map != nil {
+		encoder.Encode(m.toIntermediate())
+	} else {
+		encoder.Encode(nil)
+	}
 }
 
 // CodecDecodeSelf implements codec.Selfer.
 func (m *NodeControlDataLatestMap) CodecDecodeSelf(decoder *codec.Decoder) {
-	bm := (*LatestMap)(m)
-	bm.decoder = NodeControlDataLatestEntryDecoder
-	bm.CodecDecodeSelf(decoder)
+	out := mapRead(decoder, func(isNil bool) interface{} {
+		value := &nodeControlDataLatestEntry{}
+		if !isNil {
+			decoder.Decode(value)
+		}
+		return value
+	})
+	*m = NodeControlDataLatestMap{out}
 }
 
 // MarshalJSON shouldn't be used, use CodecEncodeSelf instead.