… report/latest_map_internal_test.go