Previously, classes like `StringLatestMap` were implemented as a shim over `LatestMap`.
This meant that the memory allocations were:
[`ps.Map` leaf] -> [`interface{}` indirection] -> `LatestEntry` -> [`interface{}` indirection] -> [string contents]

Removing `LatestMap` and storing the entry via pointer rather than value gives us instead:

[`ps.Map` leaf] -> `stringLatestEntry` -> [string contents]

Also the `LatestEntry` was copied one more time on entry into the map. Thus this PR removes 50% of the memory allocations needed to create this particular structure, per entry.

(not counting the allocation for the map key)