@bboreham 2017-03-20
Some stats:

Running the following test:

```
package report                                                                                                                                                        
                                                                                                                                                                      
import "github.com/ugorji/go/codec"                                                                                                                                   
import "io/ioutil"                                                                                                                                                    
import "testing"

func TestDecode(t *testing.T) {
    b, _ := ioutil.ReadFile("fixed.msgpack")
    for i := 0; i < 100; i++ {
        r := Report{}
        handler := &codec.MsgpackHandle{}
        d := codec.NewDecoderBytes(b, handler)
        d.Decode(&r)
    }
}
```

where `fixed.msgpack` is a 1.5MB file supplied by Adam;

On master:

```
ok  	github.com/weaveworks/scope/report	2.798s
ok  	github.com/weaveworks/scope/report	2.502s
ok  	github.com/weaveworks/scope/report	2.492s
ok  	github.com/weaveworks/scope/report	2.504s
```

With this PR:

```
ok  	github.com/weaveworks/scope/report	1.849s
ok  	github.com/weaveworks/scope/report	1.735s
ok  	github.com/weaveworks/scope/report	1.761s
ok  	github.com/weaveworks/scope/report	1.785s
```

That's about a 30% speed-up, fastest to fastest.