@2opremio 2017-03-21
Did you find generate_latest_map useful?