@bboreham 2017-03-21
Yes, it's easier to edit the code once and generate twice, than to edit it twice.